ai-model/requirements.txt
pandas
numpy 
flask 
render 
ipykernel
torch
git+https://github.com/openai/whisper.git
langchain_openai
langchain_huggingface
huggingface_hub
langchain_community
langchain
langchain-core
pydantic
transformers
accelerate
python-dotenv
torch
openai
seaborn
matplotlib
joblib

